+++ core/classifier-jobs/update.int.test.js
@@ -1,18 +1,18 @@
 const routes = require('./index')
 const models = require('../_models')
-const { migrate, truncate, expressApp, seed, seedValues, muteConsole } = require('../../common/testing/sequelize')
+const { truncate, truncateNonBase, expressApp, seedValues, muteConsole } = require('../../common/testing/sequelize')
 const request = require('supertest')
 const CLASSIFIER_JOB_STATUS = require('./classifier-job-status')
 const { WAITING, RUNNING, DONE, CANCELLED, ERROR } = require('./classifier-job-status')
 
 // Test data
-const CLASSIFIER_1 = { id: 555, name: 'sounds of the underground', version: 1, externalId: '555666', createdById: seedValues.primaryUserId, modelRunner: 'tf2', modelUrl: '???', lastExecutedAt: null, isPublic: true }
+const CLASSIFIER_1 = { id: 831, name: 'sounds of the underground', version: 1, externalId: '555666', createdById: seedValues.primaryUserId, modelRunner: 'tf2', modelUrl: '???', lastExecutedAt: null, isPublic: true }
 const CLASSIFIERS = [CLASSIFIER_1]
 
-const PROJECT_1 = { id: 'testproject1', name: 'Test project', createdById: seedValues.otherUserId }
+const PROJECT_1 = { id: 'testprojert1', name: 'Test project', createdById: seedValues.otherUserId }
 const PROJECTS = [PROJECT_1]
 
-const STREAM_1 = { id: 'LilSjZJkRK20', name: 'Test stream', start: '2021-01-02T01:00:00.000Z', end: '2021-01-02T05:00:00.000Z', isPublic: true, createdById: seedValues.otherUserId, projectId: PROJECT_1.id }
+const STREAM_1 = { id: 'LilSjZJkRK00', name: 'Test stream', start: '2021-01-02T01:00:00.000Z', end: '2021-01-02T05:00:00.000Z', isPublic: true, createdById: seedValues.otherUserId, projectId: PROJECT_1.id }
 const STREAMS = [STREAM_1]
 
 const JOB_WAITING = { id: 123, status: WAITING, classifierId: CLASSIFIER_1.id, projectId: PROJECT_1.id, queryStreams: 'Test stream, Test stream 2', queryStart: '2021-03-13', queryEnd: '2022-04-01', queryHours: '1,2', createdById: seedValues.otherUserId, created_at: '2022-06-08T08:07:49.158Z', updated_at: '2022-09-07T08:07:49.158Z', startedAt: null, completedAt: null }
@@ -22,24 +22,31 @@ const JOB_ERROR = { id: 126, status: ERROR, classifierId: CLASSIFIER_1.id, proje
 const JOB_CANCELLED = { id: 127, status: CANCELLED, classifierId: CLASSIFIER_1.id, projectId: PROJECT_1.id, queryStreams: 'Test stream, Test stream 2', queryStart: '2021-03-13', queryEnd: '2022-04-01', queryHours: '1,2', createdById: seedValues.otherUserId, created_at: '2022-06-08T08:07:49.158Z', updated_at: '2022-09-07T08:07:49.158Z', startedAt: null, completedAt: null }
 const JOBS = [JOB_WAITING, JOB_RUNNING, JOB_DONE, JOB_ERROR, JOB_CANCELLED]
 
-async function seedTestData () {
-  await models.Classifier.bulkCreate(CLASSIFIERS)
+async function seedInitialData () {
   await models.Project.bulkCreate(PROJECTS)
   await models.UserProjectRole.create({ user_id: seedValues.primaryUserId, project_id: PROJECT_1.id, role_id: seedValues.roleMember })
   await models.UserProjectRole.create({ user_id: seedValues.anotherUserId, project_id: PROJECT_1.id, role_id: seedValues.roleGuest })
   await models.Stream.bulkCreate(STREAMS)
+}
+
+async function seedTestData () {
+  await models.Classifier.bulkCreate(CLASSIFIERS)
   await models.ClassifierJob.bulkCreate(JOBS)
 }
 
 beforeAll(async () => {
-  await migrate(models.sequelize, models.Sequelize)
-  await seed(models)
   muteConsole('warn')
+  await seedInitialData()
 })
 
 beforeEach(async () => {
-  await truncate(models)
+  await truncate({ Classifier: models.Classifier, ClassifierJob: models.ClassifierJob })
   await seedTestData()
+})
+
+afterAll(async () => {
+  await truncateNonBase(models)
+  await models.sequelize.close()
 })
 
 describe('PATCH /classifier-jobs/:id', () => {
@@ -47,7 +54,6 @@ describe('PATCH /classifier-jobs/:id', () => {
   const hasPermissionApp = expressApp({ id: seedValues.otherUserId }).use('/', routes)
   const noPermissionApp = expressApp({ id: seedValues.anotherUserId }).use('/', routes)
   const superUserApp = expressApp({ is_super: true }).use('/', routes)
-  const allApps = { hasPermissionApp, noPermissionApp, superUserApp }
 
   // Split valid & invalid target status
   const { RUNNING, ...STATUS_EXCEPT_RUNNING } = CLASSIFIER_JOB_STATUS
@@ -188,13 +194,47 @@ describe('PATCH /classifier-jobs/:id', () => {
 
   describe('invalid usage', () => {
     describe('general', () => {
-      test.each(Object.entries(allApps))('400 if trying to update status with %s to RUNNING (20)', async (label, app) => {
+      test('400 if trying to update status with hasPermissionApp to RUNNING (20)', async () => {
         // Arrange
         const jobUpdate = { status: RUNNING }
 
         // Act
-        const response1 = await request(app).patch(`/${JOB_WAITING.id}`).send(jobUpdate)
-        const response2 = await request(app).patch(`/${JOB_RUNNING.id}`).send(jobUpdate)
+        const response1 = await request(hasPermissionApp).patch(`/${JOB_WAITING.id}`).send(jobUpdate)
+        const response2 = await request(hasPermissionApp).patch(`/${JOB_RUNNING.id}`).send(jobUpdate)
+
+        const jobUpdated1 = await models.ClassifierJob.findByPk(JOB_WAITING.id)
+        const jobUpdated2 = await models.ClassifierJob.findByPk(JOB_RUNNING.id)
+
+        // Assert
+        expect(response1.statusCode).toBe(400)
+        expect(response2.statusCode).toBe(400)
+        expect(jobUpdated1.status).toBe(JOB_WAITING.status)
+        expect(jobUpdated2.status).toBe(JOB_RUNNING.status)
+      })
+      test('400 if trying to update status with noPermissionApp to RUNNING (20)', async () => {
+        // Arrange
+        const jobUpdate = { status: RUNNING }
+
+        // Act
+        const response1 = await request(noPermissionApp).patch(`/${JOB_WAITING.id}`).send(jobUpdate)
+        const response2 = await request(noPermissionApp).patch(`/${JOB_RUNNING.id}`).send(jobUpdate)
+
+        const jobUpdated1 = await models.ClassifierJob.findByPk(JOB_WAITING.id)
+        const jobUpdated2 = await models.ClassifierJob.findByPk(JOB_RUNNING.id)
+
+        // Assert
+        expect(response1.statusCode).toBe(400)
+        expect(response2.statusCode).toBe(400)
+        expect(jobUpdated1.status).toBe(JOB_WAITING.status)
+        expect(jobUpdated2.status).toBe(JOB_RUNNING.status)
+      })
+      test('400 if trying to update status with superUserApp to RUNNING (20)', async () => {
+        // Arrange
+        const jobUpdate = { status: RUNNING }
+
+        // Act
+        const response1 = await request(superUserApp).patch(`/${JOB_WAITING.id}`).send(jobUpdate)
+        const response2 = await request(superUserApp).patch(`/${JOB_RUNNING.id}`).send(jobUpdate)
 
         const jobUpdated1 = await models.ClassifierJob.findByPk(JOB_WAITING.id)
         const jobUpdated2 = await models.ClassifierJob.findByPk(JOB_RUNNING.id)