+++ core/classifiers/dao/processed-segments.js
@@ -1,0 +1,9 @@
+const { ClassifierProcessedSegment } = require('../../_models')
+
+async function batchCreate (segments) {
+  await ClassifierProcessedSegment.bulkCreate(segments)
+}
+
+module.exports = {
+  batchCreate
+}