+++ core/streams/list.int.test.js
@@ -1,6 +1,7 @@
 const routes = require('.')
 const models = require('../_models')
-const { migrate, truncate, expressApp, seed, seedValues, muteConsole } = require('../../common/testing/sequelize')
+const Sequelize = require('sequelize')
+const { expressApp, seedValues, muteConsole } = require('../../common/testing/sequelize')
 const request = require('supertest')
 
 const app = expressApp()
@@ -9,17 +10,31 @@ app.use('/', routes)
 
 beforeAll(async () => {
   muteConsole()
-  await migrate(models.sequelize, models.Sequelize)
-  await seed(models)
 })
-beforeEach(async () => {
-  await truncate(models)
+
+afterEach(async () => {
+  await models.UserStreamRole.destroy({ where: {}, force: true })
+  await models.Stream.destroy({
+    where: {
+      id: {
+        [Sequelize.Op.in]: ['j123s', 'guard1', 'guard2', 'guard3', 'x456y', 'public1', 'public2', 'private2', 'p123p', 'o789o', 'AB01', 'AB02', 'AB03', 'pq1', 'pq2', 'ab1', 'ab2', 'ab3', 's1', 's2', 's3', 's4', 'o123o']
+      }
+    },
+    force: true
+  })
+  await models.UserProjectRole.destroy({ where: {}, force: true })
+  await models.Project.destroy({ where: {}, force: true })
+  await models.UserOrganizationRole.destroy({ where: {}, force: true })
+  await models.Organization.destroy({ where: {}, force: true })
+})
+
+afterAll(async () => {
+  await models.sequelize.close()
 })
 
 describe('GET /streams', () => {
   test('no results', async () => {
     const response = await request(app).get('/')
-
     expect(response.statusCode).toBe(200)
     expect(response.body).toEqual([])
   })